+++ packages/lumx-react/src/components/tooltip/Tooltip.stories.tsx
@@ -71,7 +71,7 @@ export const TooltipOnDisabledButton = () => {
 
     return (
         <>
-            <Switch checked={disabled} onToggle={setDisabled}>
+            <Switch checked={disabled} onChange={setDisabled}>
                 Toggle button disabled
             </Switch>
             <Tooltip label={'Tooltip on disabled button'}>